test: cover admission append failures
